Added call to saveName() when name is queried
It helps bot to remember more names. Moreover, users now can use the bot to save names of an arbitrary person from any chat.
Note: the call to saveName() was added after calls to handleSearch() becouse otherwise the behavoir was unpredictable: in some cases the bot was able to find newly saved name, in others it was not. It's possibly becouse redis queries are async or something like that idk